requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
Jinja2==3.1.4
python-multipart==0.0.9
gspread==6.1.2
google-auth==2.35.0
itsdangerous==2.2.0
python-dotenv==1.0.1